+++ Paper/paper.bib
@@ -2,6 +2,7 @@
   title={Python tutorial},
   author={Van Rossum, Guido and Drake Jr, Fred L},
   year={1995},
+  
   publisher={Centrum voor Wiskunde en Informatica Amsterdam, The Netherlands}
 }
 
@@ -13,6 +14,7 @@
   number={1},
   pages={117},
   year={2022},
+  doi={10.3390/nano13010117},
   publisher={MDPI}
 }
 
@@ -24,7 +26,8 @@
   number={5},
   pages={1--5},
   year={2023},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/TASC.2023.3271950}
 }
 
 @article{wei2020emerging,
@@ -35,7 +38,8 @@
   number={1},
   pages={012005},
   year={2020},
-  publisher={IOP Publishing}
+  publisher={IOP Publishing},
+  doi={10.1088/2053-1583/abc8cb}
 }
 
 
@@ -47,7 +51,8 @@
   number={4},
   pages={827--846},
   year={2021},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s42247-021-00214-5}
 }
 
 @article{azam2022laser,
@@ -58,7 +63,8 @@
   number={7},
   pages={9129--9139},
   year={2022},
-  publisher={ACS Publications}
+  publisher={ACS Publications},
+  doi={10.1021/acsanm.2c01458}
 }
 
 @article{awschalom2021quantum,
@@ -68,7 +74,9 @@
   volume={2},
   pages={1--36},
   year={2021},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/TQE.2021.3057799}
+
 }
 
 @article{tang2023spin,
@@ -78,7 +86,8 @@
   volume={1032},
   pages={1--36},
   year={2023},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.physrep.2023.09.002}
 }
 
 @article{rahman2021recent,
@@ -89,7 +98,8 @@
   number={11},
   pages={17175--17213},
   year={2021},
-  publisher={ACS Publications}
+  publisher={ACS Publications},
+  doi={10.1021/acsnano.1c06864}
 }
 
 @article{Grecco2023, 
@@ -112,7 +122,8 @@
   number={7},
   pages={548--548},
   year={1956},
-  publisher={American Institute of Physics}
+  publisher={American Institute of Physics},
+  doi={10.1063/1.1715636}
 }
 
 @article{nagaosa2010anomalous,
@@ -123,7 +134,8 @@
   number={2},
   pages={1539--1592},
   year={2010},
-  publisher={APS}
+  publisher={APS},
+  doi={10.1103/RevModPhys.82.1539}
 }
 
 @article{rojo2013review,
@@ -134,7 +146,8 @@
   number={23},
   pages={11526--11544},
   year={2013},
-  publisher={Royal Society of Chemistry}
+  publisher={Royal Society of Chemistry},
+  doi={10.1039/C3NR03242F}
 }
 
 @article{zhang2018atomic,
@@ -144,7 +157,8 @@
   volume={406},
   pages={3--17},
   year={2018},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.optcom.2017.05.015}
 }
 
 @article{li2018mechanical,
@@ -155,7 +169,8 @@
   number={5},
   pages={1701246},
   year={2018},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1002/admi.201701246}
 }
 
 @misc{ppms,
