### 2024-09-11 59f92ac
Update the mac and SR830
