+++ QuDAP/GUI/VSM/VSMDataExtraction.py
@@ -22,18 +22,19 @@ import platform
 system = platform.system()
 if system != "Windows":
     print("Not running on Windows")
-
-version_info = platform.win32_ver()
-version, build, service_pack, extra = version_info
-build_number = int(build.split('.')[2])
-if version == "10" and build_number >= 22000:
-    print("Windows 11")
-    from QuDAP.VSM.qd import *
-elif version == "10":
-    print("Windows 10")
     from VSM.qd import *
 else:
-    print("Unknown Windows version")
+    version_info = platform.win32_ver()
+    version, build, service_pack, extra = version_info
+    build_number = int(build.split('.')[2])
+    if version == "10" and build_number >= 22000:
+        print("Windows 11")
+        from QuDAP.VSM.qd import *
+    elif version == "10":
+        print("Windows 10")
+        from VSM.qd import *
+    else:
+        print("Unknown Windows version")
 
 
 try:
