+++ QuDAP/GUI/QDesign/sr830.py
@@ -27,11 +27,11 @@ class THREAD(QThread):
     def run(self):
         while self.running:
             try:
-                X = float(self.server.query("X."))  # Read the measurement result
-                Y = float(self.server.query("Y."))  # Read the measurement result
-                Mag = float(self.server.query("MAG."))  # Read the measurement result
-                Phase = float(self.server.query("PHA."))  # Read the measurement result
-                self.update_data.emit(X,Y,Mag, Phase)
+                X = float(self.server.query("OUTP? 1"))  # Read the measurement result
+                Y = float(self.server.query("OUTP? 2"))  # Read the measurement result
+                R = float(self.server.query("OUTP? 3"))  # Read the measurement result
+                Theta = float(self.server.query("OUTP? 4"))  # Read the measurement result
+                self.update_data.emit(X, Y, R, Theta)
                 time.sleep(0.01)  # Update every second
             except Exception as e:
                 print(f"Error: {e}")
@@ -49,14 +49,14 @@ class MplCanvas(FigureCanvas):
         super(MplCanvas, self).__init__(fig)
 
 
-class Lockin(QMainWindow):
+class sr830Lockin(QMainWindow):
 
     def __init__(self):
         super().__init__()
         try:
             self.isConnect = False
             self.isPlotting = False
-            self.DSP7265 = 'None'
+            self.sr830 = 'None'
             self.init_ui()
 
         except Exception as e:
@@ -90,7 +90,7 @@ class Lockin(QMainWindow):
 
 
         #  ---------------------------- PART 1 --------------------------------
-        self.current_intrument_label = QLabel("Signal Recovery 7265 DSP Lock-in Amplifier")
+        self.current_intrument_label = QLabel("Stanford Research System sr830 Lock-in Amplifier")
         self.current_intrument_label.setFont(titlefont)
         self.current_intrument_label.setStyleSheet("""
                                             QLabel{
@@ -136,8 +136,6 @@ class Lockin(QMainWindow):
         combo_text_layout.addLayout(combo_layout)
         combo_text_layout.addWidget(self.current_gpib_label, alignment=Qt.AlignmentFlag.AlignCenter)
         group_box.setLayout(combo_text_layout)
-
-
 
         #  ---------------------------- PART 3 --------------------------------
         self.lockin_reading_plotting_layout = QHBoxLayout()
@@ -185,22 +183,52 @@ class Lockin(QMainWindow):
         self.Lockin_reading_group_box.setLayout(self.lockin_reading_layout)
 
         # #  ---------------------------- PART 3 Right --------------------------------
-        #
-        # graphing_layout = QHBoxLayout()
-        # figure_group_box = QGroupBox("Graph")
-        # figure_Layout = QVBoxLayout()
-        # self.canvas = MplCanvas(self, width=5, height=4, dpi=100)
-        # toolbar = NavigationToolbar(self.canvas, self)
-        # toolbar.setStyleSheet("""
-        #                            QWidget {
-        #                                border: None;
-        #                            }
-        #                        """)
-        # figure_Layout.addWidget(toolbar, alignment=Qt.AlignmentFlag.AlignCenter)
-        # figure_Layout.addWidget(self.canvas, alignment=Qt.AlignmentFlag.AlignCenter)
-        # figure_group_box.setLayout(figure_Layout)
-        # graphing_layout.addWidget(plotting_control_group_box)
-        # graphing_layout.addWidget(figure_group_box)
+        graphing_layout = QHBoxLayout()
+        selection_Layout = QHBoxLayout()
[…]
