+++ QuDAP/GUI/QDesign/measurement.py
@@ -397,6 +397,8 @@ class Measurement(QMainWindow):
         self.ETO_radio_buttom.setFont(self.font)
         self.FMR_radio_buttom = QRadioButton("FMR")
         self.FMR_radio_buttom.setFont(self.font)
+        self.PaP_radio_buttom = QRadioButton("Pump and Probe")
+        self.PaP_radio_buttom.setFont(self.font)
         self.reset_preset_buttom = QPushButton("Reset")
         self.select_preset_buttom = QPushButton("Select")
         self.select_preset_buttom.setStyleSheet(self.Button_stylesheet)
@@ -411,6 +413,8 @@ class Measurement(QMainWindow):
         self.radio_btn_layout.addWidget(self.ETO_radio_buttom)
         self.radio_btn_layout.addStretch(1)
         self.radio_btn_layout.addWidget(self.FMR_radio_buttom)
+        self.radio_btn_layout.addStretch(2)
+        self.radio_btn_layout.addWidget(self.PaP_radio_buttom)
         self.radio_btn_layout.addStretch(2)
 
         self.select_preset_btn_layout = QHBoxLayout()
@@ -481,48 +485,103 @@ class Measurement(QMainWindow):
             self.ppms_main_layout = QHBoxLayout()
             self.connection_group_box = QGroupBox("PPMS Connection")
             self.connection_box_layout = QVBoxLayout()
+
+            self.xps_container = QWidget()
+            self.xps_main_layout = QHBoxLayout()
+            self.xps_connection_group_box = QGroupBox("XPS Connection")
+            self.xps_connection_box_layout = QVBoxLayout()
             # --------------------------------------- Part connection_PPMS ----------------------------
-            self.ppms_connection = QVBoxLayout()
-            self.ppms_host_connection_layout = QHBoxLayout()
-            self.ppms_port_connection_layout = QHBoxLayout()
-            self.ppms_connection_button_layout = QHBoxLayout()
-            self.host_label = QLabel("PPMS Host:")
-            self.host_label.setFont(self.font)
-            self.host_entry_box = QLineEdit("127.0.0.1")
-            self.host_entry_box.setFont(self.font)
-            self.host_entry_box.setFixedHeight(30)
-            self.port_label = QLabel("PPMS Port:")
-            self.port_label.setFont(self.font)
-            self.port_entry_box = QLineEdit("5000")
-            self.port_entry_box.setFont(self.font)
-            self.port_entry_box.setFixedHeight(30)
-            self.server_btn = QPushButton('Start Server')
-            self.server_btn_clicked = False
-            self.server_btn.clicked.connect(self.start_server)
-            self.connect_btn = QPushButton('Client Connect')
-            self.connect_btn.setEnabled(False)
-            self.connect_btn_clicked = False
-            self.connect_btn.clicked.connect(self.connect_client)
+            if self.PaP_radio_buttom.isChecked():
+                self.xps_connection = QVBoxLayout()
+                self.xps_host_connection_layout = QHBoxLayout()
+                self.xps_username_connection_layout = QHBoxLayout()
+                self.xps_password_connection_layout = QHBoxLayout()
+                self.xps_connection_button_layout = QHBoxLayout()
+                self.xps_host_label = QLabel("XPS Host:")
+                self.xps_host_label.setFont(self.font)
+                self.xps_host_entry_box = QLineEdit("192.168.254.254")
+                self.xps_host_entry_box.setFont(self.font)
+                self.xps_host_entry_box.setFixedHeight(30)
+                self.xps_Username_label = QLabel("Username:")
+                self.xps_Username_label.setFont(self.font)
+                self.xps_Username_entry_box = QLineEdit("Administrator")
[…]
