+++ QuDAP/requirements.txt
@@ -10,6 +10,7 @@ setuptools
 pillow
 matplotlib
 psutil
-numpy
+pythonnet
 requests
-python-pptx
+clr
+newportxps
