+++ QuDAP/GUI/QDesign/measurement.py
@@ -426,7 +426,7 @@ class Measurement(QMainWindow):
         self.Preset_group_box.setLayout(self.preset_layout)
 
         self.preset_container = QWidget()
-        self.preset_container.setFixedSize(380, 180)
+        self.preset_container.setFixedSize(380, 200)
 
         self.preset_container_layout = QHBoxLayout()
         self.preset_container_layout.addWidget(self.Preset_group_box, 1)
@@ -476,219 +476,240 @@ class Measurement(QMainWindow):
             return
 
     def preset_select(self):
-        if self.preseted == False:
-            self.preseted = True
-            with open("GUI/QSS/QComboWidget.qss", "r") as file:
-                self.QCombo_stylesheet = file.read()
-            # --------------------------------------- Part connection ----------------------------
-            self.ppms_container = QWidget()
-            self.ppms_main_layout = QHBoxLayout()
-            self.connection_group_box = QGroupBox("PPMS Connection")
-            self.connection_box_layout = QVBoxLayout()
+        try:
+            if self.preseted == False:
+                self.preseted = True
+                with open("GUI/QSS/QComboWidget.qss", "r") as file:
+                    self.QCombo_stylesheet = file.read()
+                # --------------------------------------- Part connection ----------------------------
+                self.ppms_container = QWidget()
+                self.ppms_main_layout = QHBoxLayout()
+                self.connection_group_box = QGroupBox("PPMS Connection")
+                self.connection_box_layout = QVBoxLayout()
 
-            self.xps_container = QWidget()
-            self.xps_main_layout = QHBoxLayout()
-            self.xps_connection_group_box = QGroupBox("XPS Connection")
-            self.xps_connection_box_layout = QVBoxLayout()
-            # --------------------------------------- Part connection_PPMS ----------------------------
-            if self.PaP_radio_buttom.isChecked():
-                self.xps_connection = QVBoxLayout()
-                self.xps_host_connection_layout = QHBoxLayout()
-                self.xps_username_connection_layout = QHBoxLayout()
-                self.xps_password_connection_layout = QHBoxLayout()
-                self.xps_connection_button_layout = QHBoxLayout()
-                self.xps_host_label = QLabel("XPS Host:")
-                self.xps_host_label.setFont(self.font)
-                self.xps_host_entry_box = QLineEdit("192.168.254.254")
-                self.xps_host_entry_box.setFont(self.font)
-                self.xps_host_entry_box.setFixedHeight(30)
-                self.xps_Username_label = QLabel("Username:")
-                self.xps_Username_label.setFont(self.font)
-                self.xps_Username_entry_box = QLineEdit("Administrator")
-                self.xps_Username_entry_box.setFont(self.font)
-                self.xps_Username_entry_box.setFixedHeight(30)
-                self.xps_password_label = QLabel("Password:")
-                self.xps_password_label.setFont(self.font)
-                self.xps_password_entry_box = QLineEdit("Administrator")
-                self.xps_password_entry_box.setFont(self.font)
-                self.xps_password_entry_box.setFixedHeight(30)
-                self.xps_password_entry_box.setEchoMode(QLineEdit.EchoMode.Password)
-                self.xps_server_btn = QPushButton('Start Service')
-                self.xps_server_btn_clicked = False
-                self.xps_server_btn.clicked.connect(self.xps_start_server)
-                self.xps_timeout = 10000
+                self.xps_container = QWidget()
+                self.xps_main_layout = QHBoxLayout()
+                self.xps_connection_group_box = QGroupBox("XPS Connection")
+                self.xps_connection_box_layout = QVBoxLayout()
[…]
