+++ QuDAP/GUI/QDesign/xps.py
@@ -1,0 +1,619 @@
+import time
+import platform
+from PyQt6.QtWidgets import (
+    QApplication, QWidget, QMessageBox, QGroupBox, QStackedWidget, QVBoxLayout, QLabel, QHBoxLayout
+, QCheckBox, QPushButton, QComboBox, QLineEdit)
+from PyQt6.QtGui import QFont, QIntValidator, QValidator, QDoubleValidator
+from PyQt6.QtCore import pyqtSignal, Qt, QObject, QThread
+import sys
+import pyvisa as visa
+import matplotlib
+
+matplotlib.use('QtAgg')
+from matplotlib.backends.backend_qt5agg import FigureCanvasQTAgg as FigureCanvas
+from matplotlib.backends.backend_qt5agg import NavigationToolbar2QT as NavigationToolbar
+from matplotlib.figure import Figure
+import random
+if platform.system() == 'Windows':
+    import MultiPyVu as mpv # Uncommented it on the/thesever computer
+    from MultiPyVu import MultiVuClient as mvc, MultiPyVuError
+import sys
+import traceback
+# import Data_Processing_Suite.GUI.Icon as Icon
+
+class THREAD(QThread):
+    update_data = pyqtSignal(float, str, float, str, str)  # Signal to emit the temperature and field values
+    def __init__(self, client):
+        super().__init__()
+        self.client = client
+        self.running = True
+
+    def run(self):
+        while self.running:
+            try:
+
+                T, sT = self.client.get_temperature()
+                F, sF = self.client.get_field()
+                C = self.client.get_chamber()
+                self.update_data.emit(T, sT, F, sF, C)
+                time.sleep(1)  # Update every second
+            except Exception as e:
+                self.running = False
+
+    def stop(self):
+        self.running = False
+
+class MplCanvas(FigureCanvas):
+    def __init__(self, parent=None, width=5, height=4, dpi=300):
+        fig = Figure(figsize=(width, height), dpi=dpi)
+
+        self.axes = fig.add_subplot(111)
+        super(MplCanvas, self).__init__(fig)
+
+
+class XPS(QWidget):
+
+    def __init__(self):
+        super().__init__()
+        try:
+            self.init_ui()
+            self.isConnect = False
+
+            self.server_thread = None
+        except Exception as e:
+            QMessageBox.warning(self, "Error", str(e))
+            return
+
+    def init_ui(self):
+        titlefont = QFont("Arial", 25)
+        font = QFont("Arial", 13)
+        self.setStyleSheet("background-color: white;")
+
+        # Create main vertical layout with centered alignment
+        main_layout = QVBoxLayout(self)
+        main_layout.setAlignment(Qt.AlignmentFlag.AlignCenter)
+
+        #  ---------------------------- PART 1 --------------------------------
+        self.current_intrument_label = QLabel("Newport XPS Motion Controller")
+        self.current_intrument_label.setFont(titlefont)
+        self.current_intrument_label.setStyleSheet("""
+                                            QLabel{
+                                                background-color: white;
+                                                }
+                                                """)
+        #  ---------------------------- PART 2 --------------------------------
+
+        ppms_connection = QHBoxLayout()
+        connection_group_box = QGroupBox("Device Connection")
+        # Set the layout for the group box
+        self.host_label = QLabel("Host:")
+        self.host_label.setFont(font)
+
+        self.host_entry_box = QLineEdit("127.0.0.1")
[…]
