+++ QuDAP/GUI/QDesign/measurement.py
@@ -1,4 +1,6 @@
 import sys
+import os
+import glob
 from PyQt6.QtWidgets import (
     QSizePolicy, QWidget, QMessageBox, QGroupBox, QFileDialog, QVBoxLayout, QLabel, QHBoxLayout,
     QCheckBox, QTextEdit, QPushButton, QComboBox, QLineEdit, QScrollArea, QDialog, QRadioButton, QMainWindow,
@@ -426,7 +428,7 @@ class Measurement(QMainWindow):
         self.Preset_group_box.setLayout(self.preset_layout)
 
         self.preset_container = QWidget()
-        self.preset_container.setFixedSize(380, 200)
+        self.preset_container.setFixedSize(380, 180)
 
         self.preset_container_layout = QHBoxLayout()
         self.preset_container_layout.addWidget(self.Preset_group_box, 1)
@@ -506,18 +508,12 @@ class Measurement(QMainWindow):
                         self.xps_host_entry_box.setFont(self.font)
                         self.xps_host_entry_box.setFixedHeight(30)
 
-                        self.xps_Username_label = QLabel("Username:")
-                        self.xps_Username_label.setFont(self.font)
-                        self.xps_Username_entry_box = QLineEdit("Administrator")
-                        self.xps_Username_entry_box.setFont(self.font)
-                        self.xps_Username_entry_box.setFixedHeight(30)
+                        self.xps_port_label = QLabel("Username:")
+                        self.xps_port_label.setFont(self.font)
+                        self.xps_port_entry_box = QLineEdit("5001")
+                        self.xps_port_entry_box.setFont(self.font)
+                        self.xps_port_entry_box.setFixedHeight(30)
 
-                        self.xps_password_label = QLabel("Password:")
-                        self.xps_password_label.setFont(self.font)
-                        self.xps_password_entry_box = QLineEdit("Administrator")
-                        self.xps_password_entry_box.setFont(self.font)
-                        self.xps_password_entry_box.setFixedHeight(30)
-                        self.xps_password_entry_box.setEchoMode(QLineEdit.EchoMode.Password)
 
                         self.xps_server_btn = QPushButton('Start Service')
                         self.xps_server_btn_clicked = False
@@ -526,11 +522,9 @@ class Measurement(QMainWindow):
                         self.xps_host_connection_layout.addWidget(self.xps_host_label, 1)
                         self.xps_host_connection_layout.addWidget(self.xps_host_entry_box, 2)
 
-                        self.xps_username_connection_layout.addWidget(self.xps_Username_label, 1)
-                        self.xps_username_connection_layout.addWidget(self.xps_Username_entry_box, 2)
+                        self.xps_username_connection_layout.addWidget(self.xps_port_label, 1)
+                        self.xps_username_connection_layout.addWidget(self.xps_port_entry_box, 2)
 
-                        self.xps_password_connection_layout.addWidget(self.xps_password_label, 1)
-                        self.xps_password_connection_layout.addWidget(self.xps_password_entry_box, 2)
 
                         self.xps_password_connection_layout.addStretch(1)
                         self.xps_connection_button_layout.addWidget(self.xps_server_btn)
@@ -544,7 +538,7 @@ class Measurement(QMainWindow):
 
                         self.xps_connection_group_box.setLayout(self.xps_connection)
                         self.xps_main_layout.addWidget(self.xps_connection_group_box)
-                        self.xps_container.setFixedSize(380, 200)
+                        self.xps_container.setFixedSize(380, 180)
                         self.xps_container.setLayout(self.xps_main_layout)
 
                         self.instrument_connection_layout.addWidget(self.xps_container)
@@ -593,7 +587,7 @@ class Measurement(QMainWindow):
 
[…]
