### 2024-10-15 #7194 @phibeck
Hi @arfon I could edit this one (with some caveat, see below)

Hi @CharlieGPA40 I briefly looked at the repo. Before we start, it would be good to clarify the hardware requirements for this submission. As per the documentation it seems like reviewers would need access to a very specific configuration, is that correct? According to our policy, the reviewers must be able to test the software. In order to go forward with the review, we do therefore need you to provide a way for the reviewers to run the software, for example via software emulation, or alternatively, assign reviewers with access to the hardware. Perhaps you can let me know if you can think of someone without COI who does have access, or which direction you want to proceed, thanks.
