### 2024-10-29 5dbe05d
Minor modification for initalization
