+++ QuDAP/GUI/VSM/VSMDataExtraction.py
@@ -28,10 +28,8 @@ else:
     version, build, service_pack, extra = version_info
     build_number = int(build.split('.')[2])
     if version == "10" and build_number >= 22000:
-        print("Windows 11")
         from QuDAP.VSM.qd import *
     elif version == "10":
-        print("Windows 10")
         from VSM.qd import *
     else:
         print("Unknown Windows version")
