+++ QuDAP/test2.py
@@ -1,76 +1,31 @@
+import clr
 import sys
-import csv
-from PyQt6.QtWidgets import QApplication, QMainWindow, QTableWidget, QTableWidgetItem, QVBoxLayout, QPushButton, \
-    QWidget, QFileDialog, QHeaderView
-from PyQt6.QtCore import Qt
+# import clr
+# print(dir(clr))
+
+import clr
+import glob
+import os
 
 
-class CSVTable(QMainWindow):
-    def __init__(self):
-        super().__init__()
 
-        self.setWindowTitle("CSV Viewer")
-        self.setGeometry(100, 100, 800, 600)
+directory_path = "C:/Windows/Microsoft.NET/assembly/GAC_64/Newport.XPS.CommandInterface"
+print(os.path.isdir(directory_path))
+if os.path.isdir(directory_path):
+    directories = [f for f in glob.glob(f"{directory_path}/*") if os.path.isdir(f)]
+    print(directories)
 
-        self.table_widget = QTableWidget()
-        self.table_widget.setSelectionMode(QTableWidget.SelectionMode.MultiSelection)
-        self.table_widget.setSelectionBehavior(QTableWidget.SelectionBehavior.SelectRows)
+print(f'{directories[0]}/Newport.XPS.CommandInterface.dll')
+clr.AddReference(f'{directories[0]}/Newport.XPS.CommandInterface.dll')
+# clr.AddReference("Newport.XPS.CommandInterface.dll")
+from CommandInterfaceXPS import *
 
-        load_button = QPushButton("Load CSV")
-        load_button.clicked.connect(self.load_csv)
+myXPS = XPS()
+timeout = 1000
+result = myXPS.OpenInstrument(address, port, timeout)
+if result == 0:
+    print('Open ', address, ":", port, " => Successful")
+else:
+    print('Open ', address, ":", port, " => failure ", result)
 
-        export_button = QPushButton("Export Selected Indices")
-        export_button.clicked.connect(self.export_selected_indices)
-
-        layout = QVBoxLayout()
-        layout.addWidget(load_button)
-        layout.addWidget(self.table_widget)
-        layout.addWidget(export_button)
-
-        container = QWidget()
-        container.setLayout(layout)
-        self.setCentralWidget(container)
-
-    def load_csv(self):
-        file_dialog = QFileDialog(self)
-        file_dialog.setNameFilter("CSV files (*.csv)")
-        if file_dialog.exec():
-            file_name = file_dialog.selectedFiles()[0]
-            self.populate_table(file_name)
-
-    def populate_table(self, file_name):
-        with open(file_name, newline='') as csv_file:
-            reader = csv.reader(csv_file)
-            headers = next(reader)
-            self.table_widget.setColumnCount(len(headers))
-            self.table_widget.setHorizontalHeaderLabels(headers)
-
-            self.table_widget.setRowCount(0)
-            for row_data in reader:
-                row = self.table_widget.rowCount()
-                self.table_widget.insertRow(row)
-                for column, item in enumerate(row_data):
-                    self.table_widget.setItem(row, column, QTableWidgetItem(item))
-
-        self.table_widget.resizeColumnsToContents()
-        self.table_widget.horizontalHeader().setSectionResizeMode(QHeaderView.ResizeMode.Stretch)
-
-    def export_selected_indices(self):
-        selected_rows = set()
-        selected_columns = set()
-
-        for item in self.table_widget.selectedItems():
-            selected_rows.add(item.row())
-            selected_columns.add(item.column())
-
-        print("Selected Rows:", sorted(selected_rows))
-        print("Selected Columns:", sorted(selected_columns))
-
[…]
