### 2024-11-06 #7194 @CharlieGPA40
Hi @kyleniemeyer and @phibeck, I have provided the emulation feature and sample data, which allows the software to run without hardware restrictions. However, due to the compatibility requirements of most instruments, the instrumentation functionality is limited to Windows machines. The data processing features are fully operational across all operating systems. I really appreciate the feedback you have provided!
