### 2025-01-07 #7194 @phibeck
Hi @CharlieGPA40 since the software is partially limited to Windows and (without emulation) to a specific hardware, the pool of ideal reviewers is very narrow. Do you have any suggestions for potential reviewers without COI? If not it would be helpful if you could identify a few potential reviewers from this list https://reviewers.joss.theoj.org/reviewers (without tagging them with an @), thanks.
