### 2025-02-13 #7194 @domna
Hey @phibeck, I would be up for a review. However, I currently don't have access to measurement devices and wouldn't probably be able to test run all of the capabilities of the software. I could do a general assessment of the architecture but I'm not sure if it makes sense (+ probably depends on how much I can actually run).
