### 2025-02-14 #7194 @mailhexu
Hello @phibeck @CharlieGPA40 , Thanks for the invitation. But sorry I don't think I can review this submission. While my listed expertise includes Python and magnetism, which are relevant to the topic, I'm not an experimentalist and I'm unfamiliar with the specific instrumentation involved.  Therefore, I wouldn't be able to provide a thorough and informed review.
