### 2025-02-18 #7194 @phibeck
Hi @domna,
> Hey [@phibeck](https://github.com/phibeck), I would be up for a review. However, I currently don't have access to measurement devices and wouldn't probably be able to test run all of the capabilities of the software. I could do a general assessment of the architecture but I'm not sure if it makes sense (+ probably depends on how much I can actually run).

thank you for your response. I would appreciate if you could review and test the software within the limits of your system, using the emulation feature and sample data provided by the authors. When you do so, please let me know whether there are features you cannot run and review, thank you.
