### 2025-02-18 #7194 @phibeck
Hi @lapmk

> Hi [@phibeck](https://github.com/phibeck), my available time is quite limited the next couple of weeks, but I would be willing to do the review if it is acceptable that it might take some time. After briefly reading the manuscript and checking the repository, I also have to admit that the experimental setup (hardware) looks highly specialized to me. It comprises a selection of instruments of various vendors that is in this composition apparently available in the author's lab (but the experimental techniques implemented in the software could also be performed using other instruments/vendors). Definitely, I do not have access to the required hardware. Therefore, a real test of the software with hardware communication will not be possible for me.

thank you too for your response. Since this is a very specialized tool, it is unlikely to find reviewers with access to all hardware components. In that case we ask authors to provide all necessary data to emulate the hardware, so that reviewers can run, test and review all parts of the software. I would appreciate if you could review the package within these limitations. Thanks for letting me know about your time constraints too, that is okay. I will start the review process now and send you a reminder in 3-4 weeks.
