### 2025-02-18 #7787 @editorialbot
[REVIEW]: QuDAP: an Open Source Software for Quantum Materials Characterization
**Submitting author:** @CharlieGPA40 (<a href="http://orcid.org/0000-0001-8252-3195">Chunli Tang</a>)
**Repository:** https://github.com/CharlieGPA40/QuDAP
**Branch with paper.md** (empty if default branch):
**Version:** v0.0.5
