### 2025-02-18 #7787 @CharlieGPA40
Hi Dr. Michael Krieger @lapmk, thank you for reviewing my submission. For the contribution, Jin and Samani supervised the project; Tang led the programming and PPMS measurements; White assisted with the PPMS measurements; Jia assisted with the Python coding. Let me know if you have any other questions!
