### 2025-02-19 #7787 @lapmk
Hi @CharlieGPA40. Thank you very much for your prompt reply and for the clarification. Of course, I don't want to talk into the author order, but from your explanation (and also looking at the GitHub repo) it seems to me that you have done most of the work (at least the coding). Shouldn't you be the first author on the manuscript?
