### 2025-02-20 #7787 @CharlieGPA40
Hi Dr. Michael Krieger @lapmk, thank you for your suggestion. After discussing it with all the authors and my advisors, we have updated the author list to reflect better the contributions in alignment with JOSS's software-focused nature. Please let me know if you have any further questions or suggestions.
