### 2025-03-11 577006b
Update README.md
I have implemented a fix for this issue by:

1. Ensuring pip install is not executed at runtime.
2. Adding dependency management via pyproject.toml.
3. Providing instructions for users to generate their own requirements.txt using pip-tools.
