+++ QuDAP/GUI/VSM/VSM.py
@@ -14,12 +14,8 @@ if system != "Windows":
 else:
     version_info = platform.win32_ver()
     version, build, service_pack, extra = version_info
-    build_number = int(build.split('.')[2])
-    if version == "10" and build_number >= 22000:
-        # print("Windows 11")
-        from GUI.VSM.VSMDataExtraction import *
-        from GUI.VSM.VSMDataProcessing import *
-    elif version == "10":
+    # build_number = int(build.split('.')[2])
+    if version in ["10", "11"]:
         # print("Windows 10")
         from GUI.VSM.VSMDataExtraction import *
         from GUI.VSM.VSMDataProcessing import *
