+++ QuDAP/requirements.txt
@@ -1,16 +1,0 @@
-PyQt6
-pandas
-MultiPyVu
-PyVISA
-PyVISA-sim
-numpy
-python-pptx
-pywin32
-setuptools
-pillow
-matplotlib
-psutil
-pythonnet
-requests
-clr
-newportxps
