+++ Paper/paper.bib
@@ -2,7 +2,6 @@
   title={Python tutorial},
   author={Van Rossum, Guido and Drake Jr, Fred L},
   year={1995},
-  
   publisher={Centrum voor Wiskunde en Informatica Amsterdam, The Netherlands}
 }
 
@@ -14,7 +13,6 @@
   number={1},
   pages={117},
   year={2022},
-  doi={10.3390/nano13010117},
   publisher={MDPI}
 }
 
@@ -26,8 +24,7 @@
   number={5},
   pages={1--5},
   year={2023},
-  publisher={IEEE},
-  doi={10.1109/TASC.2023.3271950}
+  publisher={IEEE}
 }
 
 @article{wei2020emerging,
@@ -38,8 +35,7 @@
   number={1},
   pages={012005},
   year={2020},
-  publisher={IOP Publishing},
-  doi={10.1088/2053-1583/abc8cb}
+  publisher={IOP Publishing}
 }
 
 
@@ -51,8 +47,7 @@
   number={4},
   pages={827--846},
   year={2021},
-  publisher={Springer},
-  doi={10.1007/s42247-021-00214-5}
+  publisher={Springer}
 }
 
 @article{azam2022laser,
@@ -63,8 +58,7 @@
   number={7},
   pages={9129--9139},
   year={2022},
-  publisher={ACS Publications},
-  doi={10.1021/acsanm.2c01458}
+  publisher={ACS Publications}
 }
 
 @article{awschalom2021quantum,
@@ -74,9 +68,7 @@
   volume={2},
   pages={1--36},
   year={2021},
-  publisher={IEEE},
-  doi={10.1109/TQE.2021.3057799}
-
+  publisher={IEEE}
 }
 
 @article{tang2023spin,
@@ -86,8 +78,7 @@
   volume={1032},
   pages={1--36},
   year={2023},
-  publisher={Elsevier},
-  doi={10.1016/j.physrep.2023.09.002}
+  publisher={Elsevier}
 }
 
 @article{rahman2021recent,
@@ -98,8 +89,7 @@
   number={11},
   pages={17175--17213},
   year={2021},
-  publisher={ACS Publications},
-  doi={10.1021/acsnano.1c06864}
+  publisher={ACS Publications}
 }
 
 @article{Grecco2023, 
@@ -122,8 +112,7 @@
   number={7},
   pages={548--548},
   year={1956},
-  publisher={American Institute of Physics},
-  doi={10.1063/1.1715636}
+  publisher={American Institute of Physics}
 }
 
 @article{nagaosa2010anomalous,
@@ -134,8 +123,7 @@
   number={2},
   pages={1539--1592},
   year={2010},
-  publisher={APS},
-  doi={10.1103/RevModPhys.82.1539}
+  publisher={APS}
 }
 
 @article{rojo2013review,
@@ -146,8 +134,7 @@
   number={23},
   pages={11526--11544},
   year={2013},
-  publisher={Royal Society of Chemistry},
-  doi={10.1039/C3NR03242F}
+  publisher={Royal Society of Chemistry}
 }
 
 @article{zhang2018atomic,
@@ -157,8 +144,7 @@
   volume={406},
   pages={3--17},
   year={2018},
-  publisher={Elsevier},
-  doi={10.1016/j.optcom.2017.05.015}
+  publisher={Elsevier}
 }
 
 @article{li2018mechanical,
@@ -169,14 +155,15 @@
   number={5},
   pages={1701246},
   year={2018},
-  publisher={Wiley Online Library},
-  doi={10.1002/admi.201701246}
+  publisher={Wiley Online Library}
 }
 
 @misc{ppms,
     title = {Quantum Design PPMS DynaCool},
     url = {https://www.qdusa.com/products/dynacool.html},
[…]
