+++ Paper/paper.md
@@ -54,13 +54,6 @@ In the following sections, we introduce the Graphic User Interface and show the 
 
 ![Block Diagram of the instrumentation of QuDAP. The QuDAP software establish the instrument communication using USB, RS232, and GPIB to control the PPMS [@ppms], BNC845 [@BNC845], Keithley 6221 [@keithley6221], Keithley 2182 [@keithley2182], and DSP7265 [@DSP7265].\label{fig1}](Figure1.png)
 
-# Instrument Descriptions
-
-The Quantum Design PPMS Dynacool offers continuous and precise field and temperature sweeping control with the vacuum system at $10^{-4}$ Torr. The temperature is tunable in the range of 1.8 - 400 K and the field can sweep from -9 T to 9 T. Equipped with the internal lock-in amplifier and current source, PPMS can characterize a plethora of magnetic properties including the hysteresis loops, coercivity, and magnetic anisotropy from vibrating sample magnetometer (VSM); effective magnetization, gyromagnetic ratio, and magnetic damping from ferromagnetic resonance (FMR); Hall effect, magnetoresistance, and I-V curves from Electrical Transport option (ETO); heat capacity from Thermal Transport Option (TTO).
-
-As summarized in Table I, our Quantum Design PPMS Dynacool is connected to Keithley Nanovoltmeter Model 2182A, Keithley Current Source 6221 from Tektronix, Berkeley Nucleonics Corp 845 microwave and RF signal generator, and 7265 DSP Lock-in Amplifier from Signal Recovery. The instruments mentioned above can be controlled by Python. The next section will discuss the implementation of Python to establish the connection to those instruments.
-![\label{Table1}](Table1.png)
-
 # Experiemnt Workflow
 
 Rotational anisotropy second harmonic generation measurements are performed using an ultrafast light source using the normal or oblique incident geometry with the fixed incident light configuration of p - or s polarization. The normal incidence geometry is when the incident light and scattered light are perpendicular to the sample surface. In contrast, the oblique incidence geometry configured both lights at an angle with respect to the normal incidence geometry. The second harmonic signal is collected by a single-photon-counting EMCCD camera. The reflected SHG intensity is recorded as a function of the azimuthal angle ϕ between the scattering plane (electric polarization) and the in-plane crystalline axis [@SHG]. 
@@ -71,13 +64,19 @@ The transport measurements were carried out using a physical property measuremen
 
 Ferromagnetic resonance spectroscopy is measured using a coplanar waveguide (CPW) and an RF diode. The microwave is injected through the CPW, and the RF diode converts the current signal into a detectable voltage. The broadband CryoFMR option, with either field modulation or Amplitude modulation, was used for FMR spectroscopy acquisition, in which a locked-in amplifier records the output signal [@tang2023spin].
 
+# Instrument Descriptions
+
+The Quantum Design PPMS Dynacool offers continuous and precise field and temperature sweeping control with the vacuum system at $10^{-4}$ Torr. The temperature is tunable in the range of 1.8 - 400 K and the field can sweep from -9 T to 9 T. Equipped with the internal lock-in amplifier and current source, PPMS can characterize a plethora of magnetic properties including the hysteresis loops, coercivity, and magnetic anisotropy from vibrating sample magnetometer (VSM); effective magnetization, gyromagnetic ratio, and magnetic damping from ferromagnetic resonance (FMR); Hall effect, magnetoresistance, and I-V curves from Electrical Transport option (ETO); heat capacity from Thermal Transport Option (TTO).
+
[…]
