+++ QuDAP/test_gui.py
@@ -1,0 +1,9 @@
+from pathlib import Path
+
+import initalization
+
+
+def test_startup(qtbot):
+
+	app = initalization.main(test_mode=True)
+	qtbot.addWidget(app)
