### 2025-06-05 #7787 @domna
Hey @phibeck, there were some answers by @CharlieGPA40 on my issues. I haven't had the time to go through them yet but will soon.
