### 2025-07-07 f7d41df
Update test_gui.py
