### 2025-07-07 0e210e0
Finished the test of opening the GUI
